refactor: restructure sip/sig11 call handling
