+++ packages/sig11/server/server.js
@@ -1,11 +1,11 @@
 import App from '@ca11/webphone/lib/app.js'
 import btoa from 'btoa'
-import Crypto from './lib/crypto.js'
-import Endpoint from './lib/endpoint.js'
+import Crypto from '../lib/crypto.js'
+import Endpoint from './endpoint.js'
 import EventEmitter from 'eventemitter3'
 import http from 'http'
 import knex from 'knex'
-import Network from './lib/network.js'
+import Network from './network.js'
 import rc from 'rc'
 import WebCrypto from 'node-webcrypto-ossl'
 import WebSocket from 'ws'
